+++ packages/appcd-core/CHANGELOG.md
@@ -1,3 +1,9 @@
+# v2.2.1
+
+ * refactor: Refactored shutdown handler to use async/await.
+ * chore: Updated telemetry config settings to latest endpoint.
+ * chore: Updated dependencies.
+
 # v2.2.0 (Mar 29, 2019)
 
  * Bumped Node.js version from 8.15.0 to 10.15.3.